requirements.txt
fastapi[standard]
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
